erApiQueueService 분리
