+++ src/er-api/er-api.consumer.ts
@@ -1,17 +1,10 @@
 import { Process, Processor } from '@nestjs/bull';
-import { InjectRepository } from '@nestjs/typeorm';
 import { Job } from 'bull';
-import { User } from 'src/entities/User.entity';
-import { ErUserInfo } from 'src/types/erApiResponse';
-import erApiService from 'src/utils/erApiService';
-import { Repository } from 'typeorm';
+import { ErApiService } from './er-api.service';
 
 @Processor('erApiQueue')
 export class ErApiConsumer {
-  constructor(
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>,
-  ) {}
+  constructor(private erService: ErApiService) {}
 
   @Process('test')
   async test(job, done) {
@@ -22,28 +15,7 @@ export class ErApiConsumer {
 
   @Process('getInfo')
   async getInfo(job: Job<getInfoParams>, done) {
-    let result: User;
-
-    const user = await this.userRepository.findOne({
-      where: { nickname: job.data.nickname },
-    });
-
-    result = user;
-
-    if (!user) {
-      const userData = await erApiService.get<ErUserInfo>(
-        `/v1/user/nickname?query=${job.data.nickname}`,
-      );
-
-      const newUser = this.userRepository.create({
-        id: userData.data.user.userNum,
-        nickname: userData.data.user.nickname,
-      });
-
-      await this.userRepository.insert(newUser);
-
-      result = newUser;
-    }
+    const result = await this.erService.getInfo(job.data.nickname);
 
     done(null, result);
   }
