+++ src/er-api/er-api.module.ts
@@ -4,13 +4,14 @@ import { ErApiConsumer } from './er-api.consumer';
 import { BullModule } from '@nestjs/bull';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from 'src/entities/User.entity';
+import { ErApiQueueService } from './er-api-queue.service';
 
 @Module({
   imports: [
     BullModule.registerQueue({ name: 'erApiQueue' }),
     TypeOrmModule.forFeature([User]),
   ],
-  providers: [ErApiService, ErApiConsumer],
-  exports: [ErApiService],
+  providers: [ErApiService, ErApiConsumer, ErApiQueueService],
+  exports: [ErApiQueueService],
 })
 export class ErApiModule {}
