+++ src/er-api/er-api.service.ts
@@ -1,33 +1,41 @@
-import { InjectQueue } from '@nestjs/bull';
 import { Injectable } from '@nestjs/common';
-import { Queue } from 'bull';
+import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/entities/User.entity';
+import { ErUserInfo } from 'src/types/erApiResponse';
+import erApiService from 'src/utils/erApiService';
+import { Repository } from 'typeorm';
 
 @Injectable()
 export class ErApiService {
   constructor(
-    @InjectQueue('erApiQueue')
-    private readonly erApiQueue: Queue,
+    @InjectRepository(User)
+    private readonly userRepository: Repository<User>,
   ) {}
 
-  async queueTest(data: string) {
-    const job = await this.erApiQueue.add(
-      'test',
-      { test: data },
-      { removeOnComplete: true, removeOnFail: true },
-    );
+  async getInfo(nickname: string): Promise<User> {
+    let result: User;
 
-    const jobResult = await job.finished();
-    return jobResult;
-  }
+    const user = await this.userRepository.findOne({
+      where: { nickname: nickname },
+    });
 
-  async getInfo(nickname: string) {
-    const job = await this.erApiQueue.add(
-      'getInfo',
-      { nickname },
-      { removeOnComplete: true, removeOnFail: true },
-    );
+    result = user;
 
-    const jobResult = await job.finished();
-    return jobResult;
+    if (!user) {
+      const userData = await erApiService.get<ErUserInfo>(
+        `/v1/user/nickname?query=${nickname}`,
+      );
+
+      const newUser = this.userRepository.create({
+        id: userData.data.user.userNum,
+        nickname: userData.data.user.nickname,
+      });
+
+      await this.userRepository.insert(newUser);
+
+      result = newUser;
+    }
+
+    return result;
   }
 }
